Fixed some algorithms

+++ RSA/src/main/java/pkg/Decrypt.java
@@ -3,11 +3,12 @@ package pkg;
 import java.math.BigInteger;
 
 public class Decrypt {
-    public static BigInteger[] Decryption(KeyPair key, BigInteger[] cryptic){
-        BigInteger[] message = new BigInteger[cryptic.length];
+    public static long[] Decryption(KeyPair key, BigInteger[] cryptic){
+
+        long[] message = new long[cryptic.length];
 
         for (int i = 0; i < message.length; i++) {
-            message[i]=cryptic[i].modPow(key.d,key.n);
+            message[i]=cryptic[i].modPow(key.d,key.n).longValue();
         }
 
         /*for (int i = 0; i < message.length; i++) {
